Add basic lights to deferred (still a bit buggy)
Lights can now be added to the scene and will get rendered with the DeferredRenderer. There are however still some things to check before this can go onto master. Biggest open questions are: where to do ambient?, how to determine which light model to use?, is the attenuation working correctly?, how do the lights look at the edges of the light model?